Moved tools/bin2c to single-level with other tools.
