--- tools/bin2c/Makefile
@@ -1,12 +1,0 @@
-.PHONY: all clean
-
-all: bin2c
-
-bin2c: bin2c.o
-	$(CC) -o $@ $<
-
-%.o: %.c
-	$(CC) -c -o $@ $<
-
-clean:
-	rm -rf bin2c bin2c.o
